Template vector assembler.
